:new: tentative IPv4 top-level framer
PCapPacketImpl does not support extracting a non-ethernet or non-sll
framed payload. Since we need to receive IPv4 packets without ethernet
or sll wrappers, a new top-level IPv4 framer is needed. This framer
checks that for the correct IP version in the first header's byte.

If packet matches, a new MacPacket is created from payload and an IPv4
packet is then extracted.